modify update program compponent
